bug fixed of join case related to pending task and party selection
