+++ frontend/micro-ui/web/micro-ui-internals/packages/modules/dristi/src/components/SelectComponentsMulti.js
@@ -174,15 +174,17 @@ const SelectComponentsMulti = ({ t, config, onSelect, formData, errors }) => {
             ></LocationComponent>
           </div>
         ))}
-      <Button
-        isDisabled={config?.disable || (config?.state && config?.state !== CaseWorkflowState.DRAFT_IN_PROGRESS)}
-        className={"add-location-btn"}
-        label={"Add Location"}
-        style={{ alignItems: "center", margin: "10px 0px" }}
-        onButtonClick={() => {
-          handleAdd();
-        }}
-      />
+      {!(config?.removeAddLocationButton === true) && (
+        <Button
+          isDisabled={config?.disable || (config?.state && config?.state !== CaseWorkflowState.DRAFT_IN_PROGRESS)}
+          className={"add-location-btn"}
+          label={"Add Location"}
+          style={{ alignItems: "center", margin: "10px 0px" }}
+          onButtonClick={() => {
+            handleAdd();
+          }}
+        />
+      )}
     </div>
   );
 };
